requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
Bootstrap-Flask==2.2.0
cachetools==7.0.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.8
colorgram.py==1.2.0
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.67.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
Markdown==3.10.2
MarkupSafe==3.0.3
multidict==6.4.4
narwhals==2.17.0
numpy==2.2.6
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
pillow==11.3.0
prettytable==3.16.0
propcache==0.3.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.3.2
PyPDF2==3.0.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.4.0
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
selenium==4.36.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spotipy==2.25.1
SQLAlchemy==2.0.25
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
trio==0.31.0
trio-websocket==0.12.2
twilio==9.6.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.0.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.20.0
